test(gateway): split oversized module tests
